2013-09-13  Sharath Kamath  <s.kamath@samsung.com>
        Redefine CLenum as unsigned long and error codes to CLint.

        Issue #220

        Reviewed by Antonio.

        As per Khronos bug[https://www.khronos.org/bugzilla/show_bug.cgi?id=958] CLenum needs to be
        unsigned long to hold OpenCL DEVICE_TYPE_ALL enum member.
        Also all the negetive members of Enum in WebCL interface will be changed to CLint.

        This also fixes the issue with getDevices(DEVICE_TYPE_ALL).

        Changing all IDL files to use CLenum, CLint, CLuint, CLboolean & CLtype
        instead of respective native types.

        cl_platform_getDevices.html in
        https://github.com/SRA-SiliconValley/webkit-webcl-conformance.

        * Modules/webcl/WebCL.idl:
        * Modules/webcl/WebCLBuffer.idl:
        * Modules/webcl/WebCLCommandQueue.idl:
        * Modules/webcl/WebCLContext.idl:
        * Modules/webcl/WebCLContextProperties.idl:
        * Modules/webcl/WebCLDevice.idl:
        * Modules/webcl/WebCLEvent.idl:
        * Modules/webcl/WebCLException.idl:
        * Modules/webcl/WebCLGL.idl:
        * Modules/webcl/WebCLGLContext.idl:
        * Modules/webcl/WebCLGLImage.idl:
        * Modules/webcl/WebCLGLObjectInfo.idl:
        * Modules/webcl/WebCLImageDescriptor.idl:
        * Modules/webcl/WebCLInputChecker.cpp: Changing all the enum values from int to unsigned long.
        (WebCore::WebCLInputChecker::isValidDeviceType):
        (WebCore::WebCLInputChecker::isValidDeviceInfoType):
        (WebCore::WebCLInputChecker::isValidMemoryObjectFlag):
        (WebCore::WebCLInputChecker::isValidAddressingMode):
        (WebCore::WebCLInputChecker::isValidFilterMode):
        (WebCore::WebCLInputChecker::isValidChannelOrder):
        (WebCore::WebCLInputChecker::isValidChannelType):
        (WebCore::WebCLInputChecker::isValidCommandQueueProperty):
        (WebCore::WebCLInputChecker::isValidGLTextureInfo):
        * Modules/webcl/WebCLInputChecker.h:
        * Modules/webcl/WebCLKernel.idl:
        * Modules/webcl/WebCLKernelArgumentTypes.idl:
        * Modules/webcl/WebCLMemoryObject.idl:
        * Modules/webcl/WebCLPlatform.idl:
        * Modules/webcl/WebCLProgram.idl:
        * Modules/webcl/WebCLSampler.idl:

+++ Source/WebCore/Modules/webcl/WebCLInputChecker.cpp
@@ -36,7 +36,7 @@
 namespace WebCore {
 namespace WebCLInputChecker {
 
-bool isValidDeviceType(long int deviceType)
+bool isValidDeviceType(unsigned long deviceType)
 {
     switch (deviceType) {
     case ComputeContext::DEVICE_TYPE_CPU:
@@ -49,7 +49,7 @@ bool isValidDeviceType(long int deviceType)
     return false;
 }
 
-bool isValidDeviceInfoType(int infoType)
+bool isValidDeviceInfoType(unsigned long infoType)
 {
     switch (infoType) {
     case ComputeContext::DEVICE_EXTENSIONS:
@@ -101,7 +101,7 @@ bool isValidDeviceInfoType(int infoType)
     return false;
 }
 
-bool isValidMemoryObjectFlag(int memoryObjectFlag)
+bool isValidMemoryObjectFlag(unsigned long memoryObjectFlag)
 {
     switch (memoryObjectFlag) {
     case ComputeContext::MEM_READ_ONLY:
@@ -113,7 +113,7 @@ bool isValidMemoryObjectFlag(int memoryObjectFlag)
     return false;
 }
 
-bool isValidAddressingMode(int value)
+bool isValidAddressingMode(unsigned long value)
 {
     switch(value) {
     case ComputeContext::ADDRESS_NONE:
@@ -127,7 +127,7 @@ bool isValidAddressingMode(int value)
     return false;
 }
 
-bool isValidFilterMode(int value)
+bool isValidFilterMode(unsigned long value)
 {
     switch(value) {
     case ComputeContext::FILTER_NEAREST:
@@ -138,7 +138,8 @@ bool isValidFilterMode(int value)
     return false;
 
 }
-bool isValidChannelOrder(int value)
+
+bool isValidChannelOrder(unsigned long value)
 {
     switch (value) {
     case ComputeContext::R:
@@ -160,7 +161,7 @@ bool isValidChannelOrder(int value)
     return false;
 }
 
-bool isValidChannelType(int value)
+bool isValidChannelType(unsigned long value)
 {
     switch (value) {
     case ComputeContext::SNORM_INT8:
@@ -184,7 +185,7 @@ bool isValidChannelType(int value)
     return false;
 }
 
-bool isValidCommandQueueProperty(int value)
+bool isValidCommandQueueProperty(unsigned long value)
 {
     switch (value) {
     case 0: // 0 as integer value CommandQueueProperty is optional.
@@ -195,7 +196,7 @@ bool isValidCommandQueueProperty(int value)
     return false;
 }
 
-bool isValidGLTextureInfo(int value)
+bool isValidGLTextureInfo(unsigned long value)
 {
     switch (value) {
     case ComputeContext::GL_TEXTURE_TARGET:
